refactor(web): split dataset file status into go/python variants
Move the Go/Python dual-track status heuristic out of shared business
code into a document-status adapter (go/python/dispatcher) behind
pickByBackend, and split ParsingStatusCell into a BackendVariant
dispatcher with go/python implementations. Go drives status from
ingestion_status (queued/progress/stopping); Python only reads the
legacy run field.
